jxl-oxide-cli: Merge binaries into one
