+++ crates/jxl-oxide-cli/src/commands/generate_fixture.rs
@@ -1,0 +1,12 @@
+use std::path::PathBuf;
+
+use clap::Parser;
+
+#[derive(Debug, Parser)]
+#[non_exhaustive]
+pub struct GenerateFixtureArgs {
+    /// Input file
+    pub input: PathBuf,
+    /// Output file
+    pub output: PathBuf,
+}
