+++ crates/jxl-oxide-cli/src/lib.rs
@@ -1,0 +1,7 @@
+pub mod commands;
+pub mod decode;
+#[cfg(feature = "__devtools")]
+pub mod generate_fixture;
+pub mod info;
+
+pub use commands::{Args, Subcommands};
